requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.1
aiohttp>=3.9.5
beautifulsoup4>=4.12.3
lxml>=5.2.1
feedparser
asyncio
playwright>=1.37.0
beautifulsoup4
curses